findbugs:change for style

+++ src/main/java/io/mycat/server/handler/ServerLoadDataInfileHandler.java
@@ -900,17 +900,15 @@ public final class ServerLoadDataInfileHandler implements LoadDataInfileHandler
             return;
         }
         File[] fileList = fileDirToDel.listFiles();
-
-        for (int i = 0; i < fileList.length; i++)
-        {
-            File file = fileList[i];
-            if (file.isFile()&&file.exists())
-            {
-                boolean delete = file.delete();
-            } else if (file.isDirectory())
-            {
-                deleteFile(file.getAbsolutePath());
-                file.delete();
+        if (fileList != null) {
+            for (int i = 0; i < fileList.length; i++) {
+                File file = fileList[i];
+                if (file.isFile() && file.exists()) {
+                    boolean delete = file.delete();
+                } else if (file.isDirectory()) {
+                    deleteFile(file.getAbsolutePath());
+                    file.delete();
+                }
             }
         }
         fileDirToDel.delete();
